Pool servers and options.

+++ src/com/oblong/jelly/pool/tcp/Factory.java
@@ -13,9 +13,9 @@ import com.oblong.jelly.PoolServer;
  */
 public class Factory implements PoolServer.Factory {
 
-    @Override public PoolServer get(String scheme, String host, short port)
+    @Override public PoolServer get(String scheme, String host, int port)
         throws PoolException {
         return null;
+        //  return new Server(new ServerAddress(host, port));
     }
-
 }
